requirement.txt
geopy==2.4.1
requests==2.32.3
pillow==11.1.0
pandas==2.2.3